+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - Nexus Prisma model and CRUD projections
 - GraphQL code generation
 - Set versions hook
+- User in GraphQL context from JWT in authorization header